Changes after 1st review

+++ include/Configuration.h
@@ -37,7 +37,7 @@ struct CHANNEL_CONFIG_T {
 struct INVERTER_CONFIG_T {
     uint64_t Serial;
     char Name[INV_MAX_NAME_STRLEN + 1];
-    uint16_t CurrentPhase;    
+    uint8_t CurrentPhase;
     CHANNEL_CONFIG_T channel[INV_MAX_CHAN_COUNT];
 };
